requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
accelerate=0.16.0=pypi_0
appdirs=1.4.4=pypi_0
blas=1.0=mkl
blis=0.7.9=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.12.7=ha878542_0
cachetools=5.3.0=pypi_0
catalogue=2.0.8=pypi_0
certifi=2022.12.7=pyhd8ed1ab_0
charset-normalizer=3.0.1=pypi_0
click=8.1.3=pypi_0
colorclass=2.2.2=pypi_0
confection=0.0.4=pypi_0
contourpy=1.0.7=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pypi_0
cymem=2.0.7=pypi_0
dataclasses=0.6=pypi_0
dbus=1.13.6=he372182_0
docker-pycreds=0.4.0=pypi_0
docopt=0.6.2=pypi_0
en-core-web-sm=3.5.0=pypi_0
expat=2.2.10=h9c3ff4c_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=pypi_0
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_1
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.16.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.51.1=pypi_0
h5py=3.8.0=pypi_0
huggingface-hub=0.12.0=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.0.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jinja2=3.1.2=pypi_0
jpeg=9b=h024ee3a_2
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
langcodes=3.3.0=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.1.0=h2733197_1
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libxcb=1.13=h7f98852_1004
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.1=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.6.3=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
murmurhash=1.0.9=pypi_0
ncurses=6.3=h5eee18b_3
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numpy=1.24.1=pypi_0
numpy-base=1.23.5=py38h31eccc5_0
oauthlib=3.2.2=pypi_0
openh264=2.1.1=h4ff587b_0
openjdk=11.0.13=h87a67e3_0
openssl=1.1.1t=h7f8727e_0
packaging=23.0=pypi_0
pandas=2.0.2=pypi_0
pathtools=0.1.2=pypi_0
pathy=0.10.1=pypi_0
pcre=8.45=h9c3ff4c_0
pillow=9.3.0=py38hace64e9_1
pip=22.3.1=py38h06a4308_0
pip-upgrader=1.4.15=pypi_0
preshed=3.0.8=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.4=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
pur=7.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.6=pypi_0
pydantic=1.10.4=pypi_0
pyparsing=3.0.9=pypi_0
python=3.8.16=h7a1cb2a_2
python-dateutil=2.8.2=pypi_0
pytorch=1.8.0=py3.8_cuda10.2_cudnn7.6.5_0
pytz=2023.3=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=h5eee18b_0
regex=2022.10.31=pypi_0
requests=2.28.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
sentry-sdk=1.19.1=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=65.6.3=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smart-open=6.3.0=pypi_0
smmap=5.0.0=pypi_0
spacy=3.5.0=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.4=pypi_0
sqlite=3.40.1=h5082296_0
srsly=2.4.5=pypi_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
terminaltables=3.1.10=pypi_0
thinc=8.1.7=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.2=pypi_0
toolz=0.12.0=pypi_0
torchaudio=0.8.0=py38
torchvision=0.9.0=py38_cu102
tqdm=4.64.1=pypi_0
transformers=4.27.0.dev0=pypi_0
typer=0.7.0=pypi_0
typing_extensions=4.4.0=py38h06a4308_0
tzdata=2023.3=pypi_0
urllib3=1.26.14=pypi_0
wandb=0.14.2=pypi_0
wasabi=1.1.1=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.10=h5eee18b_1
zipp=3.12.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.4.9=haebb681_0
